refactor files and folders
