+++ src/routes/DebtPositionCreateWizard/components/Beneficiary/BeneficiaryFieldControls/index.ts
@@ -1,0 +1,5 @@
+export { BeneficiaryControlledField } from './BeneficiaryControlledField';
+export { BeneficiaryIdentityFields } from './BeneficiaryIdentityFields';
+export { BeneficiaryAmountFields } from './BeneficiaryAmountFields';
+export { BeneficiaryPaymentFields } from './BeneficiaryPaymentFields';
+export { BeneficiaryClassificationFields } from './BeneficiaryClassificationFields';
